Moved IQuasiHttpSendOptions interface and default implementation back to Kabomu.QuasiHtttp.Client namespace
