+++ examples/Kabomu.Examples.Shared/MemoryBasedClientTransport.cs
@@ -1,4 +1,5 @@
 ﻿using Kabomu.Common;
+using Kabomu.QuasiHttp.Client;
 using Kabomu.QuasiHttp.Transport;
 using System;
 using System.Collections.Generic;
