+++ test/Kabomu.Tests.Shared/Common/ComparisonUtils.cs
@@ -2,8 +2,8 @@
 using Kabomu.Mediator.Path;
 using Kabomu.QuasiHttp;
 using Kabomu.QuasiHttp.ChunkedTransfer;
+using Kabomu.QuasiHttp.Client;
 using Kabomu.QuasiHttp.EntityBody;
-using Kabomu.QuasiHttp.Transport;
 using Newtonsoft.Json;
 using System;
 using System.Collections.Generic;
