+++ test/Kabomu.Tests.Shared/QuasiHttp/MemoryBasedClientTransport.cs
@@ -1,5 +1,6 @@
 ﻿using Kabomu.Common;
 using Kabomu.QuasiHttp;
+using Kabomu.QuasiHttp.Client;
 using Kabomu.QuasiHttp.Transport;
 using System;
 using System.Collections.Generic;
